ucs2 encoder debugging

+++ src/hypua2jamo-c/hypua2jamo.h
@@ -9,7 +9,7 @@ int hypua_p2jc_ucs2_calcsize(const unsigned short *src, int srclen);
 int hypua_p2jc_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
 int hypua_p2jd_ucs4_calcsize(const unsigned int *src, int srclen);
 int hypua_p2jd_ucs4_encode(const unsigned int *src, int srclen, unsigned int *dst);
-int hypua_p2jd_ucs2_calcsize(const unsigned int *src, int srclen);
+int hypua_p2jd_ucs2_calcsize(const unsigned short *src, int srclen);
 int hypua_p2jd_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
 
 /* decoder */
